+++ src/I2C_RTC.cpp
@@ -1,0 +1,6 @@
+// #include "I2C_RTC.h"
+
+// bool PCF8523::isTimerEnabled()
+// {
+//     return false;
+// }